requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
faiss-cpu==1.8.0.post1
numpy==1.26.4
scikit-learn==1.4.2
sentence-transformers==3.0.1
transformers==4.42.4
accelerate==0.33.0
bitsandbytes==0.43.1
peft==0.11.1
trl==0.9.6
typer==0.12.3
psutil==6.0.0
pyyaml==6.0.2
llama-cpp-python==0.2.90
click==8.1.7
